# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing digital media in archival collections.",
        "title": "Go-Medialog API",
        "contact": {},
        "version": "0.2.0"
    },
    "basePath": "/api/v0",
    "paths": {
...
        "/resources": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a list of all resources.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "List resources",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Resource"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new resource (collection) within a repository.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Create resource",
                "parameters": [
                    {
                        "description": "Resource data",
                        "name": "resource",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Resource"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Resource"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/resources/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a single resource by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Get resource",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Resource ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Resource"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes a resource by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Delete resource",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Resource ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/resources/{id}/entries": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns paginated entries for a given resource. Use all_ids=true to return only entry UUIDs.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Get resource entries",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Resource ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Return all entry IDs (no pagination)",
                        "name": "all_ids",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Results per page (default 25)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.EntryResultSet"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/resources/{id}/summary": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns media type totals and per-type summaries for a given resource.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Get resource summary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Resource ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SummaryTotalsResource"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.EntryResultSet": {
            "type": "object",
            "properties": {
                "first_page": {
                    "type": "integer"
                },
                "last_page": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Entry"
                    }
                },
                "this_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "api.SummaryTotalsResource": {
            "type": "object",
            "properties": {
                "resource_identifier": {
                    "type": "string"
                },
                "resource_title": {
                    "type": "string"
                },
                "summaries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/database.Summary"
                    }
                },
                "totals": {
                    "$ref": "#/definitions/database.Totals"
                }
            }
        },
...
        "database.Summary": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "humanSize": {
                    "type": "string"
                },
                "mediatype": {
                    "type": "string"
                },
                "size": {
                    "type": "number",
                    "format": "float64"
                }
            }
        },
        "database.Totals": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "humanSize": {
                    "type": "string"
                },
                "size": {
                    "type": "number",
                    "format": "float64"
                }
            }
        },
        "models.Accession": {
            "type": "object",
            "properties": {
                "accession_note": {
                    "description": "deprecated",
                    "type": "string"
                },
                "accession_num": {
                    "type": "string"
                },
                "accession_state": {
                    "description": "deprecated",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "resource": {
                    "$ref": "#/definitions/models.Resource"
                },
                "resource_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
        "models.Entry": {
            "type": "object",
            "properties": {
                "accession": {
                    "$ref": "#/definitions/models.Accession"
                },
                "accession_id": {
                    "type": "integer"
                },
                "box_number": {
                    "type": "string"
                },
                "content_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "description": "this should be converted to a uint",
                    "type": "integer"
                },
                "disposition_note": {
                    "type": "string"
                },
                "hdd_interface": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_filename": {
                    "type": "string"
                },
                "image_format": {
                    "type": "string"
                },
                "imaged_by": {
                    "type": "string"
                },
                "imaging_note": {
                    "type": "string"
                },
                "imaging_software": {
                    "type": "string"
                },
                "imaging_success": {
                    "type": "string"
                },
                "interface": {
                    "type": "string"
                },
                "interpretation_success": {
                    "type": "string"
                },
                "is_refreshed": {
                    "type": "boolean"
                },
                "is_transferred": {
                    "type": "boolean"
                },
                "label_text": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "manufacturer": {
                    "type": "string"
                },
                "manufacturer_serial": {
                    "type": "string"
                },
                "media_id": {
                    "type": "integer"
                },
                "media_note": {
                    "type": "string"
                },
                "mediatype": {
                    "type": "string"
                },
                "original_id": {
                    "type": "string"
                },
                "repository": {
                    "$ref": "#/definitions/models.Repository"
                },
                "repository_id": {
                    "type": "integer"
                },
                "resource": {
                    "$ref": "#/definitions/models.Resource"
                },
                "resource_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "stock_size_num": {
                    "type": "number"
                },
                "stock_unit": {
                    "type": "string"
                },
                "structure": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "description": "this should be converted to a uint",
                    "type": "integer"
                }
            }
        },
...
        "models.Repository": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
        "models.Resource": {
            "type": "object",
            "properties": {
                "collection_code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "partner_code": {
                    "type": "string"
                },
                "repository": {
                    "$ref": "#/definitions/models.Repository"
                },
                "repository_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-Medialog-Token",
            "in": "header"
        }
    }
}